+++ src/core/hle/kernel/kernel.h
@@ -253,7 +253,7 @@ private:
      */
     static const size_t MAX_COUNT = 4096;
 
-    static size_t GetSlot(Handle handle)    { return handle >> 15; }
+    static u16 GetSlot(Handle handle)    { return handle >> 15; }
     static u16 GetGeneration(Handle handle) { return handle & 0x7FFF; }
 
     /// Stores the Object referenced by the handle or null if the slot is empty.
